Add user controller, routes and auth

+++ backend/controllers/transactionController.js
@@ -1,11 +1,25 @@
 const pool = require('../config/db');
+const asyncHandler = require('express-async-handler');
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
 
 // @desc Get all transactions
-// @route GET /api/transactions
-// @access Public
+// @route GET /api/v1/transactions
+// @access Private
 exports.getTransactions = async (req, res, next) => {
+	// Get user using the id in the JWT
 	try {
-		const transactions = await pool.query(`SELECT * FROM transactions`);
+		// console.log(req.user.rows[0]);
+		const wantedUserId = req.user.rows[0].user_id;
+
+		const user = await pool.query(`SELECT * FROM users WHERE user_id='${wantedUserId}'`);
+
+		if (!user) {
+			res.status(401);
+			throw new Error('User not found');
+		}
+
+		const transactions = await pool.query(`SELECT * FROM transactions WHERE user_id='${wantedUserId}'`);
 
 		return res.status(200).json({
 			success: true,
@@ -23,21 +37,31 @@ exports.getTransactions = async (req, res, next) => {
 };
 
 // @desc Add a transaction
-// @route POST /api/transactions
-// @access Public
+// @route POST /api/v1/transactions
+// @access Private
 exports.addTransaction = async (req, res, next) => {
 	try {
-		const { text, amount } = req.body;
+		const { text, amount, category } = req.body;
 
-		// if (!text || !amount) {
-		// 	res.status(400);
-		// 	throw new Error('ValidationError');
-		// }
+		if (!text || !amount || !category) {
+			res.status(400);
+			throw new Error('ValidationError');
+		}
 
-		const transaction = await pool.query(`INSERT INTO transactions(text, amount) VALUES ($1, $2) RETURNING *`, [
-			text,
-			amount,
-		]);
+		// Get user using the id in the JWT
+		const wantedUserId = req.user.rows[0].user_id;
+		const user = await pool.query(`SELECT * FROM users WHERE user_id='${wantedUserId}'`);
+
+		if (!user) {
+			res.status(401);
+			throw new Error('User not found');
+		}
+
+		// Create transaction
+		const transaction = await pool.query(
+			`INSERT INTO transactions(user_id, text, amount, category) VALUES ($1, $2, $3, $4) RETURNING *`,
+			[wantedUserId, text, amount, category],
+		);
 
 		return res.status(201).json({ success: true, data: transaction.rows });
 	} catch (err) {
@@ -60,10 +84,20 @@ exports.addTransaction = async (req, res, next) => {
 };
 
 // @desc Delete transaction
-// @route GET /api/transactions/:id
-// @access Public
+// @route Delete /api/v1/transactions/:id
+// @access Private
 exports.deleteTransaction = async (req, res, next) => {
 	try {
+		// Get user using the id in the JWT
+		// console.log(req.user.rows[0]);
+		const wantedUserId = req.user.rows[0].user_id;
+		const user = await pool.query(`SELECT * FROM users WHERE user_id='${wantedUserId}'`);
+
+		if (!user) {
+			res.status(401);
+			throw new Error('User not found');
+		}
+
 		const transaction = await pool.query(`SELECT * FROM transactions WHERE transaction_id=${req.params.id}`);
 		const wantedTransaction = transaction.rows[0];
 
@@ -74,11 +108,18 @@ exports.deleteTransaction = async (req, res, next) => {
 			});
 		}
 
+		if (wantedTransaction.user_id !== wantedUserId) {
+			res.status(401);
+			throw new Error('Not authorized');
+		}
+
+		// Delete transaction
 		await pool.query(`DELETE FROM transactions WHERE transaction_id=${req.params.id}`);
 
 		return res.status(200).json({
 			success: true,
 			data: {},
+			message: 'Transaction deleted',
 		});
 	} catch (err) {
 		// console.log(err);
